+++ backend/src/neuroagent/agent_routine.py
@@ -67,6 +67,7 @@ class AgentsRoutine:
 
         if tools:
             create_params["parallel_tool_calls"] = agent.parallel_tool_calls
+        breakpoint()
         return await self.client.chat.completions.create(**create_params)  # type: ignore
 
     def handle_function_result(self, result: Result | Agent | BaseModel) -> Result:
@@ -244,6 +245,7 @@ class AgentsRoutine:
                     model_override=model_override,
                     stream=True,
                 )
+
                 turns += 1
                 draft_tool_calls: list[dict[str, str]] = []
                 draft_tool_calls_index = -1
@@ -347,7 +349,7 @@ class AgentsRoutine:
 
                 # Append the history with the json version
                 history.append(copy.deepcopy(message))
-
+                breakpoint()
                 # We add a true / false to check if there were tool calls.
                 message["tool_calls"] = (
                     "tool_calls" in message and message["tool_calls"]