+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Multi-model support with OpenRouter.
+
 ## [v0.6.1] - 12.06.2025
 
 ## Fixed